Significant rework of files, get for and while loops mostly workinggits!
