+++ src/fprime_gds/common/fpy/codegen.py
@@ -1,36 +1,13 @@
 from __future__ import annotations
-from abc import ABC
-import inspect
-from dataclasses import astuple, dataclass, field, fields
-from pathlib import Path
-import struct
-import traceback
-import typing
-from typing import Union, get_origin, get_args
-import zlib
-
-from lark import Transformer
+from typing import Union
 
 from fprime_gds.common.fpy.model import DirectiveErrorCode
 from fprime_gds.common.fpy.types import (
-    SPECIFIC_FLOAT_TYPES,
-    SPECIFIC_INTEGER_TYPES,
-    MACROS,
-    MAX_DIRECTIVE_SIZE,
-    MAX_DIRECTIVES_COUNT,
-    SPECIFIC_NUMERIC_TYPES,
-    SIGNED_INTEGER_TYPES,
-    UNSIGNED_INTEGER_TYPES,
     ArrayIndexType,
     CompileState,
     FieldReference,
-    FppTypeClass,
-    FpyCallable,
     FpyCmd,
     FpyMacro,
-    FpyReference,
-    FpyScope,
-    FpyTypeCtor,
     FpyVariable,
     InternalIntType,
     InternalStringType,
@@ -38,114 +15,57 @@ from fprime_gds.common.fpy.types import (
     TopDownVisitor,
     Visitor,
     convert_numeric_type,
-    create_scope,
-    get_ref_fpp_type_class,
-    is_instance_compat,
-    resolve_var,
 )
 
-from fprime_gds.common.fpy.error import CompileError
-
-# In Python 3.10+, the `|` operator creates a `types.UnionType`.
-# We need to handle this for forward compatibility, but it won't exist in 3.9.
-try:
-    from types import UnionType
-
-    UNION_TYPES = (Union, UnionType)
-except ImportError:
-    UNION_TYPES = (Union,)
-
 from fprime_gds.common.fpy.bytecode.directives import (
-    BINARY_STACK_OPS,
-    BOOLEAN_OPERATORS,
-    NUMERIC_OPERATORS,
-    UNARY_STACK_OPS,
     AllocateDirective,
     AssertDirective,
     BinaryStackOp,
     ConstCmdDirective,
     DuplicateDirective,
     FloatMultiplyDirective,
-    FloatTruncateDirective,
     GetMemberDirective,
     IntAddDirective,
     IntMultiplyDirective,
     LoadDirective,
     MemCompareDirective,
     NoOpDirective,
-    IntegerTruncate64To16Directive,
     IntegerTruncate64To32Directive,
-    IntegerTruncate64To8Directive,
-    IntegerSignedExtend16To64Directive,
-    IntegerSignedExtend32To64Directive,
-    IntegerSignedExtend8To64Directive,
     StackCmdDirective,
-    IntegerZeroExtend16To64Directive,
-    IntegerZeroExtend32To64Directive,
-    IntegerZeroExtend8To64Directive,
     Directive,
-    FloatExtendDirective,
     GotoDirective,
     IfDirective,
     NotDirective,
     PushValDirective,
-    SignedIntToFloatDirective,
     StoreConstOffsetDirective,
     StoreDirective,
     PushPrmDirective,
     PushTlmValDirective,
     UnaryStackOp,
-    UnsignedGreaterThanOrEqualDirective,
-    UnsignedIntToFloatDirective,
     UnsignedLessThanDirective,
 )
-from fprime_gds.common.loaders.ch_json_loader import ChJsonLoader
-from fprime_gds.common.loaders.cmd_json_loader import CmdJsonLoader
-from fprime_gds.common.loaders.event_json_loader import EventJsonLoader
-from fprime_gds.common.loaders.prm_json_loader import PrmJsonLoader
 from fprime_gds.common.templates.ch_template import ChTemplate
-from fprime_gds.common.templates.cmd_template import CmdTemplate
 from fprime_gds.common.templates.prm_template import PrmTemplate
-from fprime.common.models.serialize.time_type import TimeType
-from fprime.common.models.serialize.enum_type import EnumType
-from fprime.common.models.serialize.serializable_type import (
-    SerializableType as StructType,
-)
 from fprime.common.models.serialize.array_type import ArrayType
-from fprime.common.models.serialize.type_exceptions import TypeException
 from fprime.common.models.serialize.numerical_types import (
     U32Type,
-    U16Type,
     U64Type,
     U8Type,
     I64Type,
-    F32Type,
     F64Type,
-    FloatType,
-    IntegerType,
-    NumericalType,
 )
-from fprime.common.models.serialize.string_type import StringType
-from fprime.common.models.serialize.bool_type import BoolType
-from fprime_gds.common.fpy.parser import (
+from fprime_gds.common.fpy.syntax import (
     AstAssert,
     AstBinaryOp,
-    AstBoolean,
+    AstBody,
     AstElif,
     AstElifs,
     AstExpr,
     AstFor,
     AstGetAttr,
     AstGetItem,
-    AstNumber,
-    AstOp,
-    AstReference,
     AstScopedBody,
-    AstStmtWithExpr,
-    AstString,
-    Ast,
-    AstBody,
-    AstLiteral,
+    AstScopedBody,
     AstIf,
     AstAssign,
     AstFuncCall,
@@ -153,1063 +73,6 @@ from fprime_gds.common.fpy.parser import (
     AstVar,
     AstWhile,
 )
-from fprime.common.models.serialize.type_base import BaseType as FppType
-
-
-class AssignIds(TopDownVisitor):
-    """assigns a unique id to each node to allow it to be indexed in a dict"""
-
-    def __init__(self):
-        self.next_id = 0
-
-    def visit_default(self, node, state):
-        node.id = self.next_id
-        self.next_id += 1
-
-
-class SetLocalScope(Visitor):
-    def __init__(self, scope: FpyScope):
-        self.scope = scope
-
-    def visit_default(self, node: Ast, state: CompileState):
-        state.local_scopes[node] = self.scope
-
-
-class AssignLocalScopes(TopDownVisitor):
-
-    def visit_AstScopedBody(self, node: AstScopedBody, state: CompileState):
-        parent_scope = state.local_scopes.get(node)
-        # make a new scope
-        scope = FpyScope()
-        state.scope_parents[scope] = parent_scope
-        # TODO ask rob there must be a better way to do this, that isn't as slow
-        SetLocalScope(scope).run(node, state)
-
-
-class CreateVariables(TopDownVisitor):
-    """finds all variable declarations and adds them to the variable scope"""
-
-    def visit_AstAssign(self, node: AstAssign, state: CompileState):
-        if not isinstance(node.lhs, AstReference):
-            state.err("Invalid assignment", node.lhs)
-            return
-        # okay, what are we assigning to?
-        if isinstance(node.lhs, AstVar):
-            if node.type_ann is not None:
-                # new variable declaraion
-                # make sure it isn't defined in this scope
-                existing_local = state.local_scopes[node].get(node.lhs.var)
-                if existing_local is not None:
-                    # redeclaring an existing variable
-                    state.err(f"{node.lhs.var} already declared", node)
-                    return
-                # okay, declare the var
-                var = FpyVariable(node.type_ann, node)
-                # new var. put it in the table under this scope
-                state.local_scopes[node][node.lhs.var] = var
-            else:
-                # otherwise, it's a reference to an existing var
-                resolved = resolve_var(node, node.lhs.var, state)
-                if resolved is None:
-                    # unable to find this symbol
-                    state.err(
-                        f"{node.lhs.var} has not been declared",
-                        node.lhs,
-                    )
-                    return
-                # okay, we were able to resolve it
-
-        else:
-            # assigning to a member or array element. don't need to make a new variable,
-            # space already exists
-            if node.type_ann is not None:
-                # type annotation on a field assignment... it already has a type!
-                state.err("Cannot specify a type annotation for a field", node.type_ann)
-                return
-
-    def visit_AstFor(self, node: AstFor, state: CompileState):
-        # for loops have an implicit loop variable that they declare inside of their scoped body
-        existing = state.local_scopes[node.body].get(node.loop_var.var)
-
-        if existing:
-            # redeclaring an existing variable
-            # i don't think this should be possible. this var should be "declared" before
-            # anything else inside of its scope gets visited
-            assert False, (node, state.local_scopes[node.body])
-
-        var = FpyVariable(node.loop_var_type, node)
-        state.local_scopes[node.body][node.loop_var.var] = var
-        # the loop var should get resolved in the body
-        state.local_scopes[node.loop_var] = state.local_scopes[node.body]
-
-
-class CheckUseBeforeDeclare(Visitor):
-
-    def __init__(self):
-        self.currently_declared_vars: list[FpyVariable] = []
-
-    def visit_AstAssign(self, node: AstAssign, state: CompileState):
-        if not isinstance(node.lhs, AstVar):
-            # definitely not a declaration, it's a field assignment
-            return
-
-        var = resolve_var(node, node.lhs.var, state)
-
-        if var.declaration != node:
-            # this is not the node that declares this variable
-            return
-
-        # this node declares this variable
-
-        self.currently_declared_vars.append(var)
-
-    def visit_AstVar(self, node: AstVar, state: CompileState):
-        ref = state.local_scopes[node].get(node.var)
-        if ref is None:
-            # not a variable, otherwise it would be in scope. might be a type name or smth
-            return
-
-        if isinstance(ref.declaration, AstFor):
-            # declared in a for loop, will be handled as a separate pass
-            return
-        if ref.declaration.lhs == node:
-            # this is the initial name of the variable. don't crash
-            return
-
-        if ref not in self.currently_declared_vars:
-            state.err(f"'{node.var}' used before declared", node)
-            return
-
-
-class CheckUseBeforeDeclareLoopVar(TopDownVisitor):
-
-    def __init__(self):
-        self.currently_declared_vars: list[FpyVariable] = []
-
-    def visit_AstFor(self, node: AstFor, state: CompileState):
-        var = state.local_scopes[node.body][node.loop_var.var]
-
-        self.currently_declared_vars.append(var)
-
-    def visit_AstVar(self, node: AstVar, state: CompileState):
-        ref = state.local_scopes[node].get(node.var)
-        if ref is None:
-            # not a variable, otherwise it would be in scope. might be a type name or smth
-            return
-
-        if isinstance(ref.declaration, AstAssign):
-            # declared in an assignment, already checked
-            return
-        if ref.declaration.loop_var == node:
-            # this is the initial name of the variable. don't crash
-            return
-
-        if ref not in self.currently_declared_vars:
-            state.err(f"'{node.var}' used before declared", node)
-            return
-
-
-class ResolveVarsAndTypes(TopDownVisitor):
-
-    def resolve_type_reference(self, node: Ast, state: CompileState) -> bool:
-
-        # we have some special logic for types because we want them resolved early,
-        # and they are easy to resolve
-
-        def resolve(n: Ast):
-
-            if not isinstance(n, (AstVar, AstGetAttr)):
-                state.err("Unknown type", node)
-                return None
-
-            if isinstance(n, AstVar):
-                parent_scope = state.types
-                name = n.var
-            else:
-                parent_scope = resolve(n.parent)
-                name = n.attr
-
-            if parent_scope is None:
-                # error already raised
-                return None
-
-            assert isinstance(parent_scope, dict), parent_scope
-
-            node_type = parent_scope.get(name)
-            if node_type is None:
-                state.err("Unknown type", node)
-                return None
-
-            state.resolved_references[n] = node_type
-            return node_type
-
-        ret_type = resolve(node)
-        return ret_type is not None
-
-    def resolve_var_in_global_scope(
-        self,
-        node: Ast,
-        global_scope: FpyScope,
-        global_scope_name: str,
-        state: CompileState,
-    ) -> bool:
-        if not isinstance(node, AstReference):
-            return True
-
-        if not isinstance(node, AstVar):
-            return self.resolve_var_in_global_scope(
-                node.parent, global_scope, global_scope_name, state
-            )
-
-        local_scope = state.local_scopes[node]
-        resolved = None
-        while local_scope is not None and resolved is None:
-            resolved = local_scope.get(node.var)
-            local_scope = state.scope_parents[local_scope]
-
-        if resolved is None:
-            # unable to find this symbol in the hierarchy of local scopes
-            # look it up in the global scope
-            resolved = global_scope.get(node.var)
-
-        if resolved is None:
-            state.err(f"Unknown {global_scope_name}", node)
-            return False
-
-        state.resolved_references[node] = resolved
-        return True
-
-    def visit_AstFuncCall(self, node: AstFuncCall, state: CompileState):
-        if not self.resolve_var_in_global_scope(
-            node.func, state.callables, "callable", state
-        ):
-            return
-
-        for arg in node.args if node.args is not None else []:
-            # arg value refs must have values at runtime
-            if not self.resolve_var_in_global_scope(
-                arg, state.runtime_values, "value", state
-            ):
-                return
-
-    def visit_AstIf_AstElif(self, node: Union[AstIf, AstElif], state: CompileState):
-        # if condition expr refs must be "runtime values" (tlm/prm/const/etc)
-        if not self.resolve_var_in_global_scope(
-            node.condition, state.runtime_values, "value", state
-        ):
-            return
-
-    def visit_AstBinaryOp(self, node: AstBinaryOp, state: CompileState):
-        # lhs/rhs side of stack op, if they are refs, must be refs to "runtime vals"
-        if not self.resolve_var_in_global_scope(
-            node.lhs, state.runtime_values, "value", state
-        ):
-            return
-        if not self.resolve_var_in_global_scope(
-            node.rhs, state.runtime_values, "value", state
-        ):
-            return
-
-    def visit_AstUnaryOp(self, node: AstUnaryOp, state: CompileState):
-        if not self.resolve_var_in_global_scope(
-            node.val, state.runtime_values, "value", state
-        ):
-            return
-
-    def visit_AstAssign(self, node: AstAssign, state: CompileState):
-        if not self.resolve_var_in_global_scope(
-            node.lhs, state.runtime_values, "value", state
-        ):
-            return
-
-        if node.type_ann is not None:
-            # in this pass, we also go ahead and finish up the types because they're easy
-            if not self.resolve_type_reference(node.type_ann, state):
-                return
-            # okay, we know the var, we know the type, let's update the var type
-            # in the struct
-            var = state.resolved_references[node.lhs]
-            var_type = state.resolved_references[node.type_ann]
-            assert isinstance(var, FpyVariable), var
-            assert isinstance(var_type, type), var_type
-            var.type = var_type
-
-        if not self.resolve_var_in_global_scope(
-            node.rhs, state.runtime_values, "value", state
-        ):
-            return
-
-    def visit_AstFor(self, node: AstFor, state: CompileState):
-        if not self.resolve_var_in_global_scope(
-            node.loop_var, state.runtime_values, "value", state
-        ):
-            return
-        if not self.resolve_var_in_global_scope(
-            node.loop_var_type, state.types, "type", state
-        ):
-            return
-        # in this pass, we also go ahead and finish up the types because they're easy
-        if not self.resolve_type_reference(node.loop_var_type, state):
-            return
-        # okay, we know the var, we know the type, let's update the var type
-        # in the struct
-        loop_var = state.resolved_references[node.loop_var]
-        loop_var_type = state.resolved_references[node.loop_var_type]
-        assert isinstance(loop_var, FpyVariable), loop_var
-        assert isinstance(loop_var_type, type), loop_var_type
-        loop_var.type = loop_var_type
-        if not self.resolve_var_in_global_scope(
-            node.lower_bound, state.runtime_values, "value", state
-        ):
-            return
-        if not self.resolve_var_in_global_scope(
-            node.upper_bound, state.runtime_values, "value", state
-        ):
-            return
-
-    def visit_AstWhile(self, node: AstWhile, state: CompileState):
-        if not self.resolve_var_in_global_scope(
-            node.condition, state.runtime_values, "value", state
-        ):
-            return
-
-    def visit_AstAssert(self, node: AstAssert, state: CompileState):
-        if not self.resolve_var_in_global_scope(
-            node.condition, state.runtime_values, "value", state
-        ):
-            return
-        if node.exit_code is not None:
-            if not self.resolve_var_in_global_scope(
-                node.exit_code, state.runtime_values, "value", state
-            ):
-                return
-
-    def visit_AstVar(self, node: AstVar, state: CompileState):
-        # make sure that all vars are resolved when we get to them
-        # if not resolved, then the var is "outside" of a context which could resolve it
-        if node not in state.resolved_references:
-            state.err("Expression is invalid when used here", node)
-            return
-
-    def visit_AstGetItem(self, node: AstGetItem, state: CompileState):
-        if not self.resolve_var_in_global_scope(
-            node.item, state.runtime_values, "value", state
-        ):
-            return
-
-    def visit_AstLiteral_AstGetAttr(
-        self, node: Union[AstLiteral, AstGetAttr], state: CompileState
-    ):
-        # don't need to do anything for literals or getattr, but just have this here for completion's sake
-        pass
-
-    def visit_default(self, node, state):
-        # coding error, missed an expr
-        assert not is_instance_compat(node, AstStmtWithExpr), node
-
-
-class PickTypesAndResolveAttrsAndItems(Visitor):
-
-    def coerce_expr_type(
-        self, node: AstExpr, type: FppTypeClass, state: CompileState
-    ) -> bool:
-        unconverted_type = state.expr_unconverted_types[node]
-        # make sure it isn't already being coerced
-        assert unconverted_type == state.expr_converted_types[node], (unconverted_type, state.expr_converted_types[node])
-        if self.can_coerce_type(unconverted_type, type):
-            state.expr_converted_types[node] = type
-            return True
-        state.err(f"Expected {type.__name__}, found {unconverted_type.__name__}", node)
-        return False
-
-    def can_coerce_type(self, type: FppTypeClass, to_type: FppTypeClass) -> bool:
-        if type == to_type:
-            return True
-        if issubclass(type, IntegerType) and issubclass(to_type, NumericalType):
-            # we can coerce any integer into any other number
-            return True
-        if issubclass(type, FloatType) and issubclass(to_type, FloatType):
-            # we can convert any float into any float
-            return True
-        if type == InternalStringType and issubclass(to_type, StringType):
-            # we can convert the internal String type to any string type
-            return True
-
-        return False
-
-    def pick_intermediate_type(
-        self, arg_types: list[FppTypeClass], op: BinaryStackOp | UnaryStackOp
-    ) -> FppTypeClass:
-
-        if op in BOOLEAN_OPERATORS:
-            return BoolType
-
-        non_numeric = any(not issubclass(t, NumericalType) for t in arg_types)
-
-        if op == BinaryStackOp.EQUAL or op == BinaryStackOp.NOT_EQUAL:
-            if non_numeric:
-                if len(set(arg_types)) != 1:
-                    # can only compare equality between the same types
-                    return None
-                return arg_types[0]
-
-        # all arguments should be numeric
-        if non_numeric:
-            # cannot find intermediate type
-            return None
-
-        if op == BinaryStackOp.DIVIDE or op == BinaryStackOp.EXPONENT:
-            # always do true division over floats, python style
-            return F64Type
-
-        float = any(issubclass(t, FloatType) for t in arg_types)
-        unsigned = any(t in UNSIGNED_INTEGER_TYPES for t in arg_types)
-
-        if float:
-            # at least one arg is a float
-            return F64Type
-
-        if unsigned:
-            # at least one arg is unsigned
-            return U64Type
-
-        return I64Type
-
-    def is_type_constant_size(self, type: FppTypeClass) -> bool:
-        """return true if the type is statically sized"""
-        if issubclass(type, StringType):
-            return False
-
-        if issubclass(type, ArrayType):
-            return self.is_type_constant_size(type.MEMBER_TYPE)
-
-        if issubclass(type, StructType):
-            for _, arg_type, _, _ in type.MEMBER_LIST:
-                if not self.is_type_constant_size(arg_type):
-                    return False
-            return True
-
-        return True
-
-    def get_members(
-        self, node: Ast, parent_type: FppTypeClass, state: CompileState
-    ) -> list[tuple[str, FppTypeClass]] | None:
-        if not issubclass(parent_type, (StructType, TimeType)):
-            return {}
-
-        if not self.is_type_constant_size(parent_type):
-            state.err(
-                f"{parent_type} has dynamically-sized members, cannot access members",
-                node,
-            )
-            return None
-
-        member_list: list[tuple[str, FppTypeClass]] = None
-        if issubclass(parent_type, StructType):
-            member_list = [t[0:2] for t in parent_type.MEMBER_LIST]
-        else:
-            # if it is a time type, there are some "implied" members
-            member_list = []
-            member_list.append(("time_base", U16Type))
-            member_list.append(("time_context", U8Type))
-            member_list.append(("seconds", U32Type))
-            member_list.append(("useconds", U32Type))
-        return member_list
-
-    def visit_AstGetAttr(self, node: AstGetAttr, state: CompileState):
-        parent_ref = state.resolved_references.get(node.parent)
-
-        if isinstance(parent_ref, (type, FpyCallable)):
-            state.err("Unknown attribute", node)
-            return
-
-        ref = None
-        if isinstance(parent_ref, dict):
-            # getattr of a namespace
-            # parent won't actually have a type
-            ref = parent_ref.get(node.attr)
-            if ref is None:
-                state.err("Unknown attribute", node)
-                return
-        else:
-            # in all other cases, parent has at least some sort of type
-            # ref may be None (if parent is some complex expr), or it may be
-            # a tlm chan or var or etc...
-            # it may or may not have a compile time value, but it definitely has a type
-            parent_type = state.expr_unconverted_types[node.parent]
-
-            # field references store their "base reference", which is the first non-field-ref parent of
-            # the field ref. this lets you easily check what actual underlying thing (tlm chan, variable, prm)
-            # you're talking about a field of
-            base_ref = (
-                parent_ref
-                if not isinstance(parent_ref, FieldReference)
-                else parent_ref.base_ref
-            )
-            # we also calculate a "base offset" wrt. the start of the base_ref type, so you
-            # can easily pick out this field from a value of the base ref type
-            base_offset = (
-                0
-                if not isinstance(parent_ref, FieldReference)
-                else parent_ref.base_offset
-            )
-
-            member_list = self.get_members(node, parent_type, state)
-            if member_list is None:
-                return
-
-            offset = 0
-            for arg_name, arg_type in member_list:
-                if arg_name == node.attr:
-                    ref = FieldReference(
-                        is_struct_member=True,
-                        parent_expr=node.parent,
-                        type=arg_type,
-                        base_ref=base_ref,
-                        local_offset=offset,
-                        base_offset=base_offset,
-                        name=arg_name,
-                    )
-                    break
-                offset += arg_type.getMaxSize()
-                base_offset += arg_type.getMaxSize()
-
-        if ref is None:
-            state.err(
-                f"{parent_type.__name__} has no member named {node.attr}",
-                node,
-            )
-            return
-
-        ref_type = get_ref_fpp_type_class(ref)
-
-        state.resolved_references[node] = ref
-        state.expr_unconverted_types[node] = ref_type
-        state.expr_converted_types[node] = ref_type
-
-    def visit_AstGetItem(self, node: AstGetItem, state: CompileState):
-        parent_ref = state.resolved_references.get(node.parent)
-
-        if isinstance(parent_ref, (type, FpyCallable, dict)):
-            state.err("Unknown item", node)
-            return
-
-        # otherwise, we should definitely have a well-defined type for our parent expr
-
-        parent_type = state.expr_unconverted_types[node.parent]
-
-        if not self.is_type_constant_size(parent_type):
-            state.err(
-                f"{parent_type.__name__} has non-constant sized members, cannot access items",
-                node,
-            )
-            return
-
-        if not issubclass(parent_type, ArrayType):
-            state.err(f"{parent_type.__name__} is not an array", node)
-            return
-
-        # coerce the index expression to array index type
-        if not self.coerce_expr_type(node.item, ArrayIndexType, state):
-            return
-
-        base_ref = (
-            parent_ref
-            if not isinstance(parent_ref, FieldReference)
-            else parent_ref.base_ref
-        )
-
-        ref = FieldReference(
-            is_array_element=True,
-            parent_expr=node.parent,
-            type=parent_type.MEMBER_TYPE,
-            base_ref=base_ref,
-            idx_expr=node.item,
-        )
-
-        state.resolved_references[node] = ref
-        state.expr_unconverted_types[node] = parent_type.MEMBER_TYPE
-        state.expr_converted_types[node] = parent_type.MEMBER_TYPE
-
-    def visit_AstVar(self, node: AstVar, state: CompileState):
-        # already been resolved by SetScopes pass
-        ref = state.resolved_references[node]
-        if ref is None:
-            return
-        ref_type = get_ref_fpp_type_class(ref)
-
-        state.expr_unconverted_types[node] = ref_type
-        state.expr_converted_types[node] = ref_type
-
-    def visit_AstNumber(self, node: AstNumber, state: CompileState):
-        # give a best guess as to the final type of this node. we don't actually know
-        # its bitwidth or signedness yet
-        if isinstance(node.value, float):
-            result_type = F64Type
-        else:
-            result_type = InternalIntType
-
-        state.expr_unconverted_types[node] = result_type
-        state.expr_converted_types[node] = result_type
-
-    def visit_AstBinaryOp(self, node: AstBinaryOp, state: CompileState):
-        lhs_type = state.expr_unconverted_types[node.lhs]
-        rhs_type = state.expr_unconverted_types[node.rhs]
-
-        intermediate_type = self.pick_intermediate_type([lhs_type, rhs_type], node.op)
-        if intermediate_type is None:
-            state.err(
-                f"Op {node.op} undefined for {lhs_type.__name__}, {rhs_type.__name__}",
-                node,
-            )
-            return
-
-        if not self.coerce_expr_type(node.lhs, intermediate_type, state):
-            return
-        if not self.coerce_expr_type(node.rhs, intermediate_type, state):
-            return
-
-        # okay now find which actual directive we're going to use based on this intermediate
-        # type, and save it
-
-        dir = None
-        if (
-            node.op == BinaryStackOp.EQUAL or node.op == BinaryStackOp.NOT_EQUAL
-        ) and intermediate_type not in SPECIFIC_NUMERIC_TYPES:
-            dir = MemCompareDirective
-        else:
-            dir = BINARY_STACK_OPS[node.op][intermediate_type]
-
-        result_type = None
-        if node.op in NUMERIC_OPERATORS:
-            result_type = intermediate_type
-        else:
-            result_type = BoolType
-
-        state.stack_op_directives[node] = dir
-        state.expr_unconverted_types[node] = result_type
-        state.expr_converted_types[node] = result_type
-
-    def visit_AstUnaryOp(self, node: AstUnaryOp, state: CompileState):
-        val_type = state.expr_unconverted_types[node.val]
-
-        intermediate_type = self.pick_intermediate_type([val_type], node.op)
-        if intermediate_type is None:
-            state.err(f"Op {node.op} undefined for {val_type.__name__}", node)
-            return
-
-        if not self.coerce_expr_type(node.val, intermediate_type, state):
-            return
-
-        # okay now find which actual directive we're going to use based on this intermediate
-        # type, and save it
-
-        chosen_dir = UNARY_STACK_OPS[node.op][intermediate_type]
-
-        result_type = None
-        if node.op in NUMERIC_OPERATORS:
-            result_type = intermediate_type
-        else:
-            result_type = BoolType
-
-        state.stack_op_directives[node] = chosen_dir
-        state.expr_unconverted_types[node] = result_type
-        state.expr_converted_types[node] = result_type
-
-    def visit_AstString(self, node: AstString, state: CompileState):
-        state.expr_unconverted_types[node] = InternalStringType
-        state.expr_converted_types[node] = InternalStringType
-
-    def visit_AstBoolean(self, node: AstBoolean, state: CompileState):
-        state.expr_unconverted_types[node] = BoolType
-        state.expr_converted_types[node] = BoolType
-
-    def visit_AstFuncCall(self, node: AstFuncCall, state: CompileState):
-        func = state.resolved_references[node.func]
-        func_args = func.args
-        node_args = node.args if node.args else []
-
-        if len(node_args) < len(func_args):
-            state.errors.append(
-                CompileError(
-                    f"Missing arguments (expected {len(func_args)} found {len(node_args)})",
-                    node,
-                )
-            )
-            return
-        if len(node_args) > len(func_args):
-            state.errors.append(
-                CompileError(
-                    f"Too many arguments (expected {len(func_args)} found {len(node_args)})",
-                    node,
-                )
-            )
-            return
-
-        for value_expr, arg in zip(node_args, func_args):
-            arg_name, arg_type = arg
-
-            if not self.coerce_expr_type(value_expr, arg_type, state):
-                return
-
-        # got thru all args successfully
-        state.expr_unconverted_types[node] = func.return_type
-        state.expr_converted_types[node] = func.return_type
-
-    def visit_AstAssign(self, node: AstAssign, state: CompileState):
-        # should be present in resolved refs because we only let it through if
-        # variable is attr, item or var
-        lhs_ref = state.resolved_references[node.lhs]
-        if not isinstance(lhs_ref, (FpyVariable, FieldReference)):
-            state.err("Invalid assignment", node.lhs)
-            return
-
-        lhs_type = None
-        if isinstance(lhs_ref, FpyVariable):
-            lhs_type = lhs_ref.type
-        else:
-            # briefly check that we're only trying
-            # to modify an fpy var
-            if not isinstance(lhs_ref.base_ref, FpyVariable):
-                state.err("Can only assign variables", node.lhs)
-                return
-            assert (
-                state.expr_converted_types[node.lhs]
-                == state.expr_unconverted_types[node.lhs]
-            )
-            lhs_type = state.expr_converted_types[node.lhs]
-
-        # coerce the rhs into the lhs type
-        if not self.coerce_expr_type(node.rhs, lhs_type, state):
-            return
-
-    def visit_AstAssert(self, node: AstAssert, state: CompileState):
-        if not self.coerce_expr_type(node.condition, BoolType, state):
-            return
-        if node.exit_code is not None:
-            if not self.coerce_expr_type(node.exit_code, U8Type, state):
-                return
-
-    def visit_AstFor(self, node: AstFor, state: CompileState):
-        # okay we have three types, but lb gets converted to lv, so we just have lv and ub
-        # so we're going to be comparing lv to ub type, so find an intermediate
-    
-        loop_var_ref = state.resolved_references[node.loop_var]
-        loop_var_type = loop_var_ref.type
-
-        # handle the loop condition check
-        # find intermediate type. we compare two variables of loop_var_type
-        intermediate_type = self.pick_intermediate_type([loop_var_type, loop_var_type], BinaryStackOp.LESS_THAN)
-
-        if intermediate_type is None or intermediate_type is F64Type:
-            state.err(
-                f"Loop variable type must be a signed or unsigned integer type",
-                node,
-            )
-            return
-
-        state.for_loop_intermediate_type[node] = intermediate_type
-
-        # okay, and based on this intermediate type, pick a directive we're going to use to compare
-        comparison_dir = BINARY_STACK_OPS[BinaryStackOp.LESS_THAN][intermediate_type]
-
-        # store it for later
-        state.for_loop_comparison_directives[node] = comparison_dir
-
-        # upper and lower bounds must be coercible to loop variable type
-        if not self.coerce_expr_type(node.lower_bound, loop_var_type, state):
-            return
-        if not self.coerce_expr_type(node.upper_bound, loop_var_type, state):
-            return
-
-        # handle increment loop var
-        # this looks like:
-        # loop_var = loop_var + 1
-        # LOAD, EXTEND, PUSH U64(1), IADD, TRUNC, STORE
-
-        self._visit(AstAssign(node.meta, node.loop_var, AstBinaryOp(node.meta, node.loop_var, "+", AstNumber(node.meta, 1))), state)
-
-    def visit_AstWhile(self, node: AstWhile, state: CompileState):
-        if not self.coerce_expr_type(node.condition, BoolType, state):
-            return
-
-    def visit_AstIf_AstElif(self, node: Union[AstIf, AstElif], state: CompileState):
-        if not self.coerce_expr_type(node.condition, BoolType, state):
-            return
-
-    def visit_default(self, node, state):
-        # coding error, missed an expr
-        assert not is_instance_compat(node, AstStmtWithExpr), node
-
-class LowerForLoopSyntax(Transformer):
-    """rewrites the tree to expand for loops into their given components"""
-    def visit_AstFor(self, node: AstFor, state: CompileState):
-
-
-
-
-
-class AllocateVariables(Visitor):
-    def visit_AstAssign(self, node: AstAssign, state: CompileState):
-        lhs_ref = state.resolved_references[node.lhs]
-        if not isinstance(lhs_ref, FpyVariable):
-            # it's a field ref, ignore it. don't need any more space for it
-            return
-
-        assert lhs_ref is not None
-        assert lhs_ref.type is not None
-
-        value_size = lhs_ref.type.getMaxSize()
-
-        if lhs_ref.lvar_offset is None:
-            # doesn't have an lvar idx, allocate one
-            lvar_offset = state.lvar_array_size_bytes
-            state.lvar_array_size_bytes += value_size
-            lhs_ref.lvar_offset = lvar_offset
-
-    def visit_AstFor(self, node: AstFor, state: CompileState):
-        # allocate space for the loop var
-        loop_var_ref = state.resolved_references[node.loop_var]
-        assert isinstance(loop_var_ref, FpyVariable)
-        lvar_offset = state.lvar_array_size_bytes
-        state.lvar_array_size_bytes += loop_var_ref.type.getMaxSize()
-        loop_var_ref.lvar_offset = lvar_offset
-
-        # allocate space for the upper bound var
-        # type of ub var is intermediate type
-        upper_bound_type = state.for_loop_intermediate_type[node]
-        lvar_offset = state.lvar_array_size_bytes
-        upper_bound_var = FpyVariable(None, node, upper_bound_type, lvar_offset)
-        state.lvar_array_size_bytes += upper_bound_var.type.getMaxSize()
-        upper_bound_var.lvar_offset = lvar_offset
-        # store ub var in a dict for later
-        state.for_loop_upper_bound_variables[node] = upper_bound_var
-
-
-class CalculateConstExprValues(Visitor):
-    """for each expr, try to calculate its constant value and store it in a map. stores None if no value could be
-    calculated at compile time, and NothingType if the expr had no value"""
-
-    def const_coerce_type(
-        self, from_val: FppType, to_type: FppTypeClass, node: Ast, state: CompileState
-    ) -> FppType | None:
-        try:
-            if type(from_val) == to_type:
-                return from_val
-            if issubclass(to_type, StringType):
-                assert type(from_val) == InternalStringType, type(from_val)
-                return to_type(from_val.val)
-            if issubclass(to_type, FloatType):
-                assert issubclass(type(from_val), NumericalType), type(from_val)
-                return to_type(float(from_val.val))
-            if issubclass(to_type, IntegerType):
-                assert issubclass(type(from_val), IntegerType), type(from_val)
-                return to_type(int(from_val.val))
-            assert False, (from_val, type(from_val), to_type)
-        except TypeException as e:
-            state.err(f"For type {type(from_val).__name__}: {e}", node)
-            return None
-
-    def visit_AstLiteral(self, node: AstLiteral, state: CompileState):
-        uncoerced_type = state.expr_unconverted_types[node]
-
-        try:
-            expr_value = uncoerced_type(node.value)
-        except TypeException as e:
-            state.err(f"For type {uncoerced_type.__name__}: {e}", node)
-            return
-
-        coerced_type = state.expr_converted_types[node]
-        if coerced_type != uncoerced_type:
-            expr_value = self.const_coerce_type(expr_value, coerced_type, node, state)
-            if expr_value is None:
-                return
-
-        state.expr_converted_values[node] = expr_value
-
-    def visit_AstGetAttr(self, node: AstGetAttr, state: CompileState):
-
-        unconverted_type = state.expr_unconverted_types[node]
-        converted_type = state.expr_converted_types[node]
-        ref = state.resolved_references[node]
-        expr_value = None
-        if isinstance(ref, (type, dict, FpyCallable)):
-            # these types have no value
-            state.expr_converted_values[node] = NothingType()
-            assert unconverted_type == converted_type, (
-                unconverted_type,
-                converted_type,
-            )
-            return
-        elif isinstance(ref, (ChTemplate, PrmTemplate, FpyVariable)):
-            # has a value but won't try to calc at compile time
-            state.expr_converted_values[node] = None
-            return
-        elif isinstance(ref, FppType):
-            expr_value = ref
-        elif isinstance(ref, FieldReference):
-            parent_value = state.expr_converted_values[node.parent]
-            if parent_value is None:
-                # no compile time constant value for our parent here
-                state.expr_converted_values[node] = None
-                return
-
-            # we are accessing an attribute of something with an fprime value at compile time
-            # we must be getting a member
-            if isinstance(parent_value, StructType):
-                expr_value = parent_value._val[node.attr]
-            elif isinstance(parent_value, TimeType):
-                if node.attr == "seconds":
-                    expr_value = U32Type(parent_value.seconds)
-                elif node.attr == "useconds":
-                    expr_value = U32Type(parent_value.useconds)
-                elif node.attr == "time_base":
-                    expr_value = U16Type(parent_value.timeBase)
-                elif node.attr == "time_context":
-                    expr_value = U8Type(parent_value.timeContext)
-                else:
-                    assert False, node.attr
-            else:
-                assert False, parent_value
-
-        assert expr_value is not None
-
-        assert isinstance(expr_value, unconverted_type), (expr_value, unconverted_type)
-
-        if converted_type != unconverted_type:
-            expr_value = self.const_coerce_type(expr_value, converted_type, node, state)
-            if expr_value is None:
-                return
-        state.expr_converted_values[node] = expr_value
-
-    def visit_AstGetItem(self, node: AstGetItem, state: CompileState):
-        ref = state.resolved_references[node]
-        # get item can only be a field reference
-        assert isinstance(ref, FieldReference), ref
-
-        parent_value = state.expr_converted_values[node.parent]
-
-        if parent_value is None:
-            # no compile time constant value for our parent here
-            state.expr_converted_values[node] = None
-            return
-
-        assert isinstance(parent_value, ArrayType), parent_value
-
-        idx = state.expr_converted_values.get(node.item)
-        if idx is None:
-            # no compile time constant value for our index
-            state.expr_converted_values[node] = None
-            return
-
-        assert isinstance(idx, U64Type)
-
-        expr_value = parent_value._val[idx._val]
-
-        uncoerced_type = state.expr_unconverted_types[node]
-        assert isinstance(expr_value, uncoerced_type), (expr_value, uncoerced_type)
-
-        coerced_type = state.expr_converted_types[node]
-        if coerced_type != uncoerced_type:
-            expr_value = self.const_coerce_type(expr_value, coerced_type, node, state)
-            if expr_value is None:
-                return
-        state.expr_converted_values[node] = expr_value
-
-    def visit_AstVar(self, node: AstVar, state: CompileState):
-        unconverted_type = state.expr_unconverted_types[node]
-        converted_type = state.expr_converted_types[node]
-        ref = state.resolved_references[node]
-        expr_value = None
-        if isinstance(ref, (type, dict, FpyCallable)):
-            # these types have no value
-            state.expr_converted_values[node] = NothingType()
-            assert unconverted_type == converted_type, (
-                unconverted_type,
-                converted_type,
-            )
-            return
-        elif isinstance(ref, (ChTemplate, PrmTemplate, FpyVariable)):
-            # has a value but won't try to calc at compile time
-            state.expr_converted_values[node] = None
-            return
-        elif isinstance(ref, FppType):
-            expr_value = ref
-        elif isinstance(ref, FieldReference):
-            assert False, ref
-
-        assert expr_value is not None
-
-        assert isinstance(expr_value, unconverted_type), (expr_value, unconverted_type)
-
-        if converted_type != unconverted_type:
-            expr_value = self.const_coerce_type(expr_value, converted_type, node, state)
-            if expr_value is None:
-                return
-        state.expr_converted_values[node] = expr_value
-
-    def visit_AstFuncCall(self, node: AstFuncCall, state: CompileState):
-        func = state.resolved_references[node.func]
-        assert isinstance(func, FpyCallable)
-        # gather arg values
-        arg_values = [
-            state.expr_converted_values[e]
-            for e in (node.args if node.args is not None else [])
-        ]
-        unknown_value = any(v for v in arg_values if v is None)
-        if unknown_value:
-            # we will have to calculate this at runtime
-            state.expr_converted_values[node] = None
-            return
-
-        expr_value = None
-
-        if isinstance(func, FpyTypeCtor):
-            # actually construct the type
-            if issubclass(func.type, StructType):
-                instance = func.type()
-                # pass in args as a dict
-                # t[0] is the arg name
-                arg_dict = {t[0]: v for t, v in zip(func.type.MEMBER_LIST, arg_values)}
-                instance._val = arg_dict
-                expr_value = instance
-
-            elif issubclass(func.type, ArrayType):
-                instance = func.type()
-                instance._val = arg_values
-                expr_value = instance
-
-            elif func.type == TimeType:
-                expr_value = TimeType(*[val.val for val in arg_values])
-
-            else:
-                # no other FppTypeClasses have ctors
-                assert False, func.return_type
-        else:
-            # don't try to calculate the value of this function call
-            # it's something like a cmd or macro
-            state.expr_converted_values[node] = None
-            return
-
-        uncoerced_type = state.expr_unconverted_types[node]
-        assert isinstance(expr_value, uncoerced_type), (expr_value, uncoerced_type)
-
-        coerced_type = state.expr_converted_types[node]
-        if coerced_type != uncoerced_type:
-            expr_value = self.const_coerce_type(expr_value, coerced_type, node, state)
-            if expr_value is None:
-                return
-
-        state.expr_converted_values[node] = expr_value
-
-    def visit_AstOp(self, node: AstOp, state: CompileState):
-        # we do not calculate compile time value of operators at the moment
-        state.expr_converted_values[node] = None
-
-    def visit_default(self, node, state):
-        # coding error, missed an expr
-        assert not is_instance_compat(node, AstExpr), node
-
 
 class GenerateConstExprDirectives(Visitor):
     """for each expr with a constant compile time value, generate
@@ -1638,7 +501,6 @@ class GenerateExprMacrosAndCmds(Visitor):
         # convert the lower bound into the intermediate type
         pass
 
-# TODO rewrite the AST here
 
 class CountNodeDirectives(Visitor):
     """count the number of directives that will be generated by each node"""
@@ -1661,6 +523,19 @@ class CountNodeDirectives(Visitor):
 
         state.node_dir_counts[node] = count
 
+    def visit_AstWhile(self, node: AstWhile, state: CompileState):
+        count = 0
+        # include the condition
+        count += state.node_dir_counts[node.condition]
+        # include if stmt
+        count += 1
+        # include body
+        count += state.node_dir_counts[node.body]
+        # include a goto begin of while
+        count += 1
+        state.node_dir_counts[node] = count
+
+
     def visit_AstElifs(self, node: AstElifs, state: CompileState):
         count = 0
         for case in node.cases:
@@ -1681,29 +556,9 @@ class CountNodeDirectives(Visitor):
 
         state.node_dir_counts[node] = count
 
-    def visit_AstFor(self, node: AstFor, state: CompileState):
-        count = 0
-        # include upper bound push
-        count += state.node_dir_counts[node.upper_bound]
-        # include push val and store ub
-        count += 2
-
-        # include lb push
-        count += state.node_dir_counts[node.lower_bound]
-        # include push val and store lb
-        count += 2
-        # include end of loop check: load lv, load ub, cmp, if
-        count += 4
-        # include body
-        count += state.node_dir_counts[node.body]
-        # include increment lv: load lv, push 1, add, push lvar offset, store
-        count += 5
-        # include goto loop check
-        count += 1
-
-        state.node_dir_counts[node] = count
-
-    def visit_AstBody(self, node: Union[AstBody, AstScopedBody], state: CompileState):
+    def visit_AstBody(
+        self, node: Union[AstScopedBody, AstBody], state: CompileState
+    ):
         count = 0
         if (
             isinstance(node, AstScopedBody)
@@ -1727,7 +582,9 @@ class CalculateStartLineIdx(TopDownVisitor):
     """based on the number of directives generated by each node, calculate the start line idx
     of each node's directives"""
 
-    def visit_AstBody(self, node: Union[AstBody, AstScopedBody], state: CompileState):
+    def visit_AstBody(
+        self, node: Union[AstScopedBody, AstBody], state: CompileState
+    ):
         if node not in state.start_line_idx:
             state.start_line_idx[node] = 0
 
@@ -1741,6 +598,17 @@ class CalculateStartLineIdx(TopDownVisitor):
         for stmt in node.stmts:
             state.start_line_idx[stmt] = line_idx
             line_idx += state.node_dir_counts[stmt]
+
+    def visit_AstWhile(self, node: AstWhile, state: CompileState):
+        line_idx = state.start_line_idx[node]
+        state.start_line_idx[node.condition] = line_idx
+        line_idx += state.node_dir_counts[node.condition]
+        # include if stmt
+        line_idx += 1
+        state.start_line_idx[node.body] = line_idx
+        line_idx += state.node_dir_counts[node.body]
+        # include goto stmt begin
+        line_idx += 1
 
     def visit_AstIf(self, node: AstIf, state: CompileState):
         line_idx = state.start_line_idx[node]
@@ -1824,101 +692,31 @@ class GenerateBodyDirectives(Visitor):
 
         state.directives[node] = all_dirs
 
-    def visit_AstFor(self, node: AstFor, state: CompileState):
+    def visit_AstWhile(self, node: AstWhile, state: CompileState):
         start_line_idx = state.start_line_idx[node]
-        # we need to do a few things:
-        # 1. initialize the loop var, and store the lower bound in it
-        # 2. calculate the upper bound and store it in the upper bound var
-        
-        # then, on each loop:
-        # 3. check loop condition
-        # 4. execute body
-        # 5. increment loop var
-        # 6. go back to start
-
-        # 1. set loop var to lower bound
-        # push lower bound to stack
-        # the converted type of these dirs should be the loop_var type
-        dirs = state.directives[node.lower_bound].copy()
-        # now store in lvar
-        loop_var = state.resolved_references[node.loop_var]
-        assert state.expr_converted_types[node.lower_bound] == loop_var.type
-        # store in loop var
-        dirs.append(
-            StoreConstOffsetDirective(loop_var.lvar_offset, loop_var.type.getMaxSize())
-        )
-
-        # 2. calc upper bound and store it in upper bound var
-
-        # push upper bound to stack
-        # the converted type of the ub expr should be intermediate_type
-        dirs.extend(state.directives[node.upper_bound])
-        intermediate_type = state.for_loop_intermediate_type[node]
-        upper_bound_var = state.for_loop_upper_bound_variables[node]
-        # store in upper bound var
-        dirs.append(
-            StoreConstOffsetDirective(
-                upper_bound_var.lvar_offset, intermediate_type.getMaxSize()
-            )
-        )
-
-
-        # okay, loop and UB vars have the right initial value
-        # begin the loop!
-
-        # 3. now add the "end-of-loop" check
-
-        end_of_loop_check_start_idx = start_line_idx + len(dirs)
-
-        upper_bound_var = state.for_loop_upper_bound_variables[node]
-
-        # loop var should be "loop var" type
-        lhs_cmp_dirs = state.directives[node.loop_var]
-        # ub should be intermediate type
-        rhs_cmp_dirs = [LoadDirective(upper_bound_var.lvar_offset, upper_bound_var.type.getMaxSize())]
-
-        # convert lv to intermediate type
-        lhs_cmp_dirs.extend(convert_numeric_type(loop_var.type, intermediate_type))
-
-        # which variant of the op did we pick?
-        cmp_dir = state.for_loop_comparison_directives[node]
-
-        # push lhs and rhs to stack
-        dirs.extend(lhs_cmp_dirs)
-        dirs.extend(rhs_cmp_dirs)
-        # generate the actual cmp op itself
-        dirs.append(cmp_dir())
-
+        # push the condition to the stack
+        dirs = state.directives[node.condition]
+        # if the cond is true, fall thru, otherwise go to end
         if_dir = IfDirective(-1)
         dirs.append(if_dir)
-        # okay now include body
+        # run body
         dirs.extend(state.directives[node.body])
-        # okay increment loop var
-        # push loop var to stack
-        dirs.append(LoadDirective(loop_var.lvar_offset, loop_var.type.getMaxSize()))
-        # convert loop var to intermediate type
-        dirs.extend(convert_numeric_type(loop_var.type, intermediate_type))
-        # push 1 to stack
-        dirs.append(PushValDirective(U64Type(1).serialize()))
-        # add them
-        dirs.append(IntAddDirective())
-        # convert back into loop var type
-        dirs.extend(convert_numeric_type(intermediate_type, loop_var.type))
-        # store in lvar array
-        dirs.append(
-            StoreConstOffsetDirective(loop_var.lvar_offset, loop_var.type.getMaxSize())
-        )
-        # okay, done with this iteration of the loop. go back up to the end-of-loop check
-        dirs.append(GotoDirective(end_of_loop_check_start_idx))
-
-        # and now update the if directive to go to just past the end of the body
-        if_dir.false_goto_dir_index = start_line_idx + len(dirs)
-
-        # okay! all done
+        # go back to condition check
+        dirs.append(GotoDirective(start_line_idx))
+        end_line_idx = start_line_idx + len(dirs)
+        if_dir.false_goto_dir_index = end_line_idx
 
         state.directives[node] = dirs
 
-    def visit_AstBody(self, node: Union[AstBody, AstScopedBody], state: CompileState):
+
+    def visit_AstFor(self, node: AstFor, state: CompileState):
+        # should have been desugared out
+        assert False, node
+
+
+    def visit_AstBody(
+        self, node: Union[AstScopedBody, AstBody], state: CompileState
+    ):
         dirs = []
         if (
             isinstance(node, AstScopedBody)
@@ -1933,144 +731,3 @@ class GenerateBodyDirectives(Visitor):
 
         state.directives[node] = dirs
 
-
-def get_base_compile_state(dictionary: str) -> CompileState:
-    """return the initial state of the compiler, based on the given dict path"""
-    cmd_json_dict_loader = CmdJsonLoader(dictionary)
-    (cmd_id_dict, cmd_name_dict, versions) = cmd_json_dict_loader.construct_dicts(
-        dictionary
-    )
-
-    ch_json_dict_loader = ChJsonLoader(dictionary)
-    (ch_id_dict, ch_name_dict, versions) = ch_json_dict_loader.construct_dicts(
-        dictionary
-    )
-    prm_json_dict_loader = PrmJsonLoader(dictionary)
-    (prm_id_dict, prm_name_dict, versions) = prm_json_dict_loader.construct_dicts(
-        dictionary
-    )
-    event_json_dict_loader = EventJsonLoader(dictionary)
-    (event_id_dict, event_name_dict, versions) = event_json_dict_loader.construct_dicts(
-        dictionary
-    )
-    # the type name dict is a mapping of a fully qualified name to an fprime type
-    # here we put into it all types found while parsing all cmds, params and tlm channels
-    type_name_dict: dict[str, FppTypeClass] = cmd_json_dict_loader.parsed_types
-    type_name_dict.update(ch_json_dict_loader.parsed_types)
-    type_name_dict.update(prm_json_dict_loader.parsed_types)
-    type_name_dict.update(event_json_dict_loader.parsed_types)
-
-    # enum const dict is a dict of fully qualified enum const name (like Ref.Choice.ONE) to its fprime value
-    enum_const_name_dict: dict[str, FppType] = {}
-
-    # find each enum type, and put each of its values in the enum const dict
-    for name, typ in type_name_dict.items():
-        if issubclass(typ, EnumType):
-            for enum_const_name, val in typ.ENUM_DICT.items():
-                enum_const_name_dict[name + "." + enum_const_name] = typ(
-                    enum_const_name
-                )
-
-    # insert the implicit types into the dict
-    type_name_dict["Fw.Time"] = TimeType
-    for typ in SPECIFIC_NUMERIC_TYPES:
-        type_name_dict[typ.get_canonical_name()] = typ
-    type_name_dict["bool"] = BoolType
-    # note no string type at the moment
-
-    cmd_response_type = type_name_dict["Fw.CmdResponse"]
-    callable_name_dict: dict[str, FpyCallable] = {}
-    # add all cmds to the callable dict
-    for name, cmd in cmd_name_dict.items():
-        cmd: CmdTemplate
-        args = []
-        for arg_name, _, arg_type in cmd.arguments:
-            args.append((arg_name, arg_type))
-        # cmds are thought of as callables with a Fw.CmdResponse return value
-        callable_name_dict[name] = FpyCmd(cmd_response_type, args, cmd)
-
-    # for each type in the dict, if it has a constructor, create an FpyTypeCtor
-    # object to track the constructor and put it in the callable name dict
-    for name, typ in type_name_dict.items():
-        args = []
-        if issubclass(typ, StructType):
-            for arg_name, arg_type, _, _ in typ.MEMBER_LIST:
-                args.append((arg_name, arg_type))
-        elif issubclass(typ, ArrayType):
-            for i in range(0, typ.LENGTH):
-                args.append(("e" + str(i), typ.MEMBER_TYPE))
-        elif issubclass(typ, TimeType):
-            args.append(("time_base", U16Type))
-            args.append(("time_context", U8Type))
-            args.append(("seconds", U32Type))
-            args.append(("useconds", U32Type))
-        else:
-            # bool, enum, string or numeric type
-            # none of these have callable ctors
-            continue
-
-        callable_name_dict[name] = FpyTypeCtor(typ, args, typ)
-
-    # for each macro function, add it to the callable dict
-    for macro_name, macro in MACROS.items():
-        callable_name_dict[macro_name] = macro
-
-    state = CompileState(
-        tlms=create_scope(ch_name_dict),
-        prms=create_scope(prm_name_dict),
-        types=create_scope(type_name_dict),
-        callables=create_scope(callable_name_dict),
-        consts=create_scope(enum_const_name_dict),
-    )
-    return state
-
-
-def compile(body: AstScopedBody, dictionary: str) -> list[Directive] | CompileError:
-    state = get_base_compile_state(dictionary)
-    passes: list[Visitor] = [
-        AssignIds(),
-        AssignLocalScopes(),
-        # based on assignment syntax nodes, we know which variables exist where
-        CreateVariables(),
-        CheckUseBeforeDeclare(),
-        CheckUseBeforeDeclareLoopVar(),
-        # now that variables have been defined, all names/attributes/indices (references)
-        # should be defined
-        ResolveVarsAndTypes(),
-        # now that we know what all refs point to, we should be able to figure out the type
-        # of every expression
-        PickTypesAndResolveAttrsAndItems(),
-        # now that expr types have been narrowed down, we can allocate lvar space for variables
-        AllocateVariables(),
-        # okay, now that we're sure we're passing in all the right args to each func,
-        # we can calculate values of type ctors etc etc
-        CalculateConstExprValues(),
-        # for expressions which have constant values, generate corresponding directives
-        # to put the expr on the stack
-        GenerateConstExprDirectives(),
-        # generate directives to calculate exprs, macros and cmds at runtime and put them
-        # on the stack
-        GenerateExprMacrosAndCmds(),
-        # count the number of directives generated by each node
-        CountNodeDirectives(),
-        # calculate the index that the node will correspond to in the output file
-        CalculateStartLineIdx(),
-        # generate directives for each body node, including the root
-        GenerateBodyDirectives(),
-    ]
-
-    for compile_pass in passes:
-        compile_pass.run(body, state)
-        if len(state.errors) != 0:
-            return state.errors[0]
-
-    dirs = state.directives[body]
-    if len(dirs) > MAX_DIRECTIVES_COUNT:
-        err = CompileError(
-            f"Too many directives in sequence (expected less than {MAX_DIRECTIVES_COUNT}, had {len(dirs)})"
-        )
-        return err
-
-    # TODO check lvar array not > max stack size (AND TEST THIS!)
-
-    return dirs
