+++ src/fprime_gds/common/fpy/grammar.lark
@@ -22,8 +22,8 @@ assign: _expr [":" _expr] "=" _expr
 
 
 # loops
-for_stmt.1: "for" var ":" _expr "in" _expr "to" _expr ":" scoped_body
-while_stmt: "while" _expr ":" scoped_body
+for_stmt.1: "for" var ":" _expr "from" _expr "to" _expr ":" body
+while_stmt: "while" _expr ":" body
 
 # branching
