+++ src/fprime_gds/common/fpy/test_helpers.py
@@ -6,7 +6,6 @@ from fprime_gds.common.fpy.types import deserialize_directives, serialize_direct
 from fprime_gds.common.fpy.model import DirectiveErrorCode, FpySequencerModel
 from fprime_gds.common.fpy.bytecode.directives import Directive
 from fprime_gds.common.fpy.main import compile_main
-from fprime_gds.common.fpy.parser import parse
 from fprime_gds.common.loaders.ch_json_loader import ChJsonLoader
 from fprime_gds.common.loaders.cmd_json_loader import CmdJsonLoader
 from fprime_gds.common.loaders.event_json_loader import EventJsonLoader
